Move kernel tests to miden-tx
