+++ bench-tx/src/main.rs
@@ -17,8 +17,8 @@ use miden_objects::{
     Felt,
 };
 use miden_tx::{
-    host::BasicAuthenticator, utils::Serializable, MockDataStore, TransactionExecutor,
-    TransactionHost, TransactionProgress,
+    host::BasicAuthenticator, testing::data_store::MockDataStore, utils::Serializable,
+    TransactionExecutor, TransactionHost, TransactionProgress,
 };
 use rand::rngs::StdRng;
 use vm_processor::{ExecutionOptions, RecAdviceProvider, Word};
