+++ miden-tx/src/kernel_tests/test_prologue.rs
@@ -1,41 +1,39 @@
 use alloc::{collections::BTreeMap, vec::Vec};
 
+use miden_lib::transaction::{
+    memory::{
+        MemoryOffset, ACCT_CODE_ROOT_PTR, ACCT_DB_ROOT_PTR, ACCT_ID_AND_NONCE_PTR, ACCT_ID_PTR,
+        ACCT_STORAGE_ROOT_PTR, ACCT_STORAGE_SLOT_TYPE_DATA_OFFSET, ACCT_VAULT_ROOT_PTR,
+        BATCH_ROOT_PTR, BLK_HASH_PTR, BLOCK_METADATA_PTR, BLOCK_NUMBER_IDX,
+        CHAIN_MMR_NUM_LEAVES_PTR, CHAIN_MMR_PEAKS_PTR, CHAIN_ROOT_PTR, CONSUMED_NOTE_ARGS_OFFSET,
+        CONSUMED_NOTE_ASSETS_HASH_OFFSET, CONSUMED_NOTE_ASSETS_OFFSET, CONSUMED_NOTE_ID_OFFSET,
+        CONSUMED_NOTE_INPUTS_HASH_OFFSET, CONSUMED_NOTE_METADATA_OFFSET,
+        CONSUMED_NOTE_NUM_ASSETS_OFFSET, CONSUMED_NOTE_NUM_INPUTS_OFFSET,
+        CONSUMED_NOTE_SCRIPT_ROOT_OFFSET, CONSUMED_NOTE_SECTION_OFFSET,
+        CONSUMED_NOTE_SERIAL_NUM_OFFSET, INIT_ACCT_HASH_PTR, INIT_NONCE_PTR, NOTE_ROOT_PTR,
+        NULLIFIER_COMMITMENT_PTR, NULLIFIER_DB_ROOT_PTR, PREV_BLOCK_HASH_PTR, PROOF_HASH_PTR,
+        PROTOCOL_VERSION_IDX, TIMESTAMP_IDX, TX_SCRIPT_ROOT_PTR,
+    },
+    TransactionKernel,
+};
 use miden_objects::{
     assembly::ProgramAst,
     testing::{
         account::MockAccountType,
+        notes::AssetPreservationStatus,
         storage::{generate_account_seed, AccountSeedType},
     },
     transaction::{PreparedTransaction, TransactionArgs, TransactionScript},
     Digest,
 };
-use miden_tx::host::testing::{
+use miden_tx::testing::{
+    mock_inputs, mock_inputs_with_account_seed,
     utils::{consumed_note_data_ptr, prepare_transaction, run_tx, run_tx_with_inputs},
     MockHost,
 };
 use vm_processor::AdviceInputs;
 
 use super::{build_module_path, ContextId, Felt, Process, ProcessState, Word, TX_KERNEL_DIR, ZERO};
-use crate::{
-    testing::{mock_inputs, mock_inputs_with_account_seed, notes::AssetPreservationStatus},
-    transaction::{
-        memory::{
-            MemoryOffset, ACCT_CODE_ROOT_PTR, ACCT_DB_ROOT_PTR, ACCT_ID_AND_NONCE_PTR, ACCT_ID_PTR,
-            ACCT_STORAGE_ROOT_PTR, ACCT_STORAGE_SLOT_TYPE_DATA_OFFSET, ACCT_VAULT_ROOT_PTR,
-            BATCH_ROOT_PTR, BLK_HASH_PTR, BLOCK_METADATA_PTR, BLOCK_NUMBER_IDX,
-            CHAIN_MMR_NUM_LEAVES_PTR, CHAIN_MMR_PEAKS_PTR, CHAIN_ROOT_PTR,
-            CONSUMED_NOTE_ARGS_OFFSET, CONSUMED_NOTE_ASSETS_HASH_OFFSET,
-            CONSUMED_NOTE_ASSETS_OFFSET, CONSUMED_NOTE_ID_OFFSET, CONSUMED_NOTE_INPUTS_HASH_OFFSET,
-            CONSUMED_NOTE_METADATA_OFFSET, CONSUMED_NOTE_NUM_ASSETS_OFFSET,
-            CONSUMED_NOTE_NUM_INPUTS_OFFSET, CONSUMED_NOTE_SCRIPT_ROOT_OFFSET,
-            CONSUMED_NOTE_SECTION_OFFSET, CONSUMED_NOTE_SERIAL_NUM_OFFSET, INIT_ACCT_HASH_PTR,
-            INIT_NONCE_PTR, NOTE_ROOT_PTR, NULLIFIER_COMMITMENT_PTR, NULLIFIER_DB_ROOT_PTR,
-            PREV_BLOCK_HASH_PTR, PROOF_HASH_PTR, PROTOCOL_VERSION_IDX, TIMESTAMP_IDX,
-            TX_SCRIPT_ROOT_PTR,
-        },
-        TransactionKernel,
-    },
-};
 
 const PROLOGUE_FILE: &str = "prologue.masm";
