+++ miden-tx/tests/integration/scripts/faucet.rs
@@ -18,7 +18,7 @@ use miden_objects::{
     transaction::TransactionArgs,
     Felt, Word, ZERO,
 };
-use miden_tx::{MockDataStore, TransactionExecutor};
+use miden_tx::{testing::data_store::MockDataStore, TransactionExecutor};
 use rand_chacha::{rand_core::SeedableRng, ChaCha20Rng};
 
 use crate::{
